refactor: moves sort alphabetically to lib
