+++ src/views/product-tutorials-view/helpers/index.ts
@@ -18,18 +18,3 @@ export function filterCollections(
 		(c: ClientCollection) => c.theme === theme && !c.slug.includes('onboarding')
 	)
 }
-
-export function sortAlphabetically<A, B>(property: string) {
-	return (a: A, b: B) => {
-		const A = a[property].toUpperCase()
-		const B = b[property].toUpperCase()
-
-		if (A < B) {
-			return -1
-		}
-		if (A > B) {
-			return 1
-		}
-		return 0
-	}
-}
